Lvl5 Vent and Vent should need to be walked into
I redid how vents worked, now the vents are in a separate tilemap gameObject than Walls1. The vent map has the vent layer attributed to them, and in the project settings, under the physics 2D section I make the default and vents layer ignore collision with each other.

This solves the vent in level 5 and the player can simply walk through the vent.