Lvl 17 - Both pressure plates need to have weight on them for door to open
On WeightedPressurePad_JoaoBeijinho script I created a public bool and WeightedPressurePad_JoaoBeijinho to reference the script on the other pressure pad.

On OnTriggerEnter2D toggle the pressure pad as active, and if the other pressure pad is active, open the door.

On OnTriggerExit2D toggle the pressure pad as inactive and close the door.

I removed the other door for the script, as only one is meant to be open using this pressure pads.

In the inspector of each pressure pad, I dragged the other pressure pad to this pressure pad inspector, on the place of the Other Pressure Pad.