Death While holding box - Respawn with box
In pushableObject_JoaoBeijinho Script, created a reference to the PlayerHealthHunger_MarioFernandes Script.

Created two separate functions for the Grab and UnGrab that was being done in the if statements in the Update().

In update, since in the PlayerHealth script the Die() function and the current health variables are private, I used the HealthSlider.

If the HealthSlider reaches 1 or below, call the UnGrab function.